학사 대시보드 오류수정
1. 이수 학점에서 소수점으로 값을 받아오고, 소수점 값을 입력할 수 없었던 문제를 수정

+++ src/main/java/com/webservice/graduate_coach/service/AcademyService.java
@@ -161,8 +161,8 @@ public class AcademyService {
         params.put("seongyo_req_credit", graduate.getOptionalEdu());
         params.put("jeontam_req_credit", graduate.getFoundationMajor());
         params.put("advanced_req_credit", graduate.getTotalLevel());
-        params.put("jeonpil_req_credit", earn_major.getReqMajorCredit());
-        params.put("jeonseon_req_credit", earn_major.getOptMajorCredit());
+        params.put("jeonpil_req_credit", earn_major.getReqMajorCredit().intValue());
+        params.put("jeonseon_req_credit", earn_major.getOptMajorCredit().intValue());
         params.put("foreign_cert_list", foreign_certs);
         params.put("comm_cert_list", comm_certs);
